+++ nutanix/config.go
@@ -9,6 +9,7 @@ import (
 	"github.com/terraform-providers/terraform-provider-nutanix/nutanix/sdks/v3/foundation"
 	"github.com/terraform-providers/terraform-provider-nutanix/nutanix/sdks/v3/karbon"
 	v3 "github.com/terraform-providers/terraform-provider-nutanix/nutanix/sdks/v3/prism"
+	"github.com/terraform-providers/terraform-provider-nutanix/nutanix/sdks/v3/selfservice"
 	"github.com/terraform-providers/terraform-provider-nutanix/nutanix/sdks/v4/clusters"
 	"github.com/terraform-providers/terraform-provider-nutanix/nutanix/sdks/v4/datapolicies"
 	"github.com/terraform-providers/terraform-provider-nutanix/nutanix/sdks/v4/dataprotection"
@@ -121,6 +122,10 @@ func (c *Config) Client() (*Client, error) {
 	if err != nil {
 		return nil, err
 	}
+	calmClient, err := selfservice.NewCalmClient(configCreds)
+	if err != nil {
+		return nil, err
+	}
 	return &Client{
 		WaitTimeout:         c.WaitTimeout,
 		API:                 v3Client,
@@ -138,6 +143,7 @@ func (c *Config) Client() (*Client, error) {
 		VmmAPI:              vmmClient,
 		DataPoliciesAPI:     dataPoliciesClient,
 		LcmAPI:              LcmClient,
+		CalmAPI:             calmClient,
 	}, nil
 }
 
@@ -159,4 +165,5 @@ type Client struct {
 	VmmAPI              *vmm.Client
 	DataPoliciesAPI     *datapolicies.Client
 	LcmAPI              *lcm.Client
+	CalmAPI             *selfservice.Client
 }