+++ nutanix/services/iamv2/datasource_nutanix_users_v2_test.go
@@ -83,6 +83,22 @@ func TestAccV2NutanixUsersDatasource_WithLimit(t *testing.T) {
 	})
 }
 
+func TestAccV2NutanixUsersDatasource_WithInvalidFilter(t *testing.T) {
+	resource.Test(t, resource.TestCase{
+		PreCheck:     func() { acc.TestAccPreCheck(t) },
+		Providers:    acc.TestAccProviders,
+		CheckDestroy: testAccCheckNutanixUserDestroy,
+		Steps: []resource.TestStep{
+			{
+				Config: testUsersDatasourceV4WithInvalidFilterConfig(),
+				Check: resource.ComposeTestCheckFunc(
+					resource.TestCheckResourceAttr(datasourceNameUsers, "users.#", "0"),
+				),
+			},
+		},
+	})
+}
+
 func testUsersDatasourceV4Config(filepath, name string) string {
 	return fmt.Sprintf(`
 		locals{
@@ -171,6 +187,7 @@ func testUsersDatasourceV4WithLimitConfig(filepath, name string) string {
 		}
 	`, filepath, name)
 }
+
 
 func TestAccV2NutanixUsersDataSourceServiceAccount(t *testing.T) {
 	r := acctest.RandInt()
@@ -267,4 +284,11 @@ func checkServiceAccountValues(expectedKeys []map[string]string, resourceName st
 		}
 		return nil
 	}
+
+func testUsersDatasourceV4WithInvalidFilterConfig() string {
+	return `
+	data "nutanix_users_v2" "test" {
+		filter = "username eq 'invalid'"
+	}
+	`
 }